+++ src/devices/MettlerToledo/MettlerToledoDevice.cs
@@ -4,11 +4,10 @@
 
 using System;
 using System.Text;
-using System.Device.Gpio;
-using System.Device.I2c;
-using System.Device.Spi;
 using System.IO.Ports;
 using UnitsNet;
+using Iot.Device.MettlerToledo.Readings;
+using Iot.Device.MettlerToledo.Exceptions;
 
 namespace Iot.Device.MettlerToledo
 {
@@ -18,6 +17,7 @@ namespace Iot.Device.MettlerToledo
     public class MettlerToledoDevice : IDisposable
     {
         private SerialPort _serialPort;
+        private bool _shouldDisposeSerialPort = true;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="MettlerToledoDevice"/> class.
@@ -27,6 +27,19 @@ namespace Iot.Device.MettlerToledo
             _serialPort = new SerialPort(port, baudrate, Parity.None);
             _serialPort.Encoding = Encoding.ASCII;
             _serialPort.NewLine = "\r\n";
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="MettlerToledoDevice"/> class.
+        /// </summary>
+        /// <param name="serialPort">Serial port to use</param>
+        /// <param name="shouldDispose">Should the serial port be disposed by this class?</param>
+        public MettlerToledoDevice(SerialPort serialPort, bool shouldDispose = true)
+        {
+            _serialPort = serialPort;
+            _serialPort.Encoding = Encoding.ASCII;
+            _serialPort.NewLine = "\r\n";
+            _shouldDisposeSerialPort = shouldDispose;
         }
 
         /// <summary>
@@ -44,6 +57,7 @@ namespace Iot.Device.MettlerToledo
         /// <summary>
         /// Opens a connection to the MT-SACS scale.
         /// </summary>
+        /// <exception cref="UnauthorizedAccessException"></exception>
         public void Open()
         {
             _serialPort.Open();
@@ -54,12 +68,18 @@ namespace Iot.Device.MettlerToledo
         /// </summary>
         public void Close()
         {
-            _serialPort.Close();
+            if (_serialPort.IsOpen)
+            {
+                _serialPort.Close();
+            }
         }
 
         /// <summary>
         /// Zeroes the balance/scale. If <paramref name="immediately"/> is set to true, will balance regardless of stability.
         /// </summary>
+        /// <exception cref="CommandNotExecutableException"></exception>
+        /// <exception cref="BalanceOutOfRangeException"></exception>
+        /// <exception cref="UnknownResultException"></exception>
         /// <returns>Boolean indicating whether re-zero was performed under stable conditions</returns>
         public bool Zero(bool immediately = false)
         {
@@ -80,7 +100,7 @@ namespace Iot.Device.MettlerToledo
                 case "D": // force non-stable success
                     return false;
                 default:
-                    throw new Exception("Unknown result received");
+                    throw new UnknownResultException("Unknown result received");
             }
         }
 
@@ -93,12 +113,13 @@ namespace Iot.Device.MettlerToledo
         /// <item>If balance (scale) is on standby, it is switched on</item>
         /// </list>
         /// </summary>
+        /// <exception cref="UnknownResultException"></exception>
         public void Reset()
         {
             var response = SendCommandWithResponse("@");
             if (response[0] != "I4")
             {
-                throw new Exception("Unknown result received");
+                throw new UnknownResultException("Unknown result received");
             }
         }
 
@@ -106,6 +127,8 @@ namespace Iot.Device.MettlerToledo
         /// Writes text to the balance display
         /// </summary>
         /// <param name="text">Text to write to the display</param>
+        /// <exception cref="UnknownResultException"></exception>
+        /// <exception cref="CommandNotExecutableException"></exception>
         public void WriteToDisplay(string text)
         {
             if (text.Contains("\""))
@@ -124,21 +147,25 @@ namespace Iot.Device.MettlerToledo
                 case "L":
                     throw new CommandNotExecutableException(command); // todo: maybe show a different error?
                 default:
-                    throw new Exception("Unknown result received");
+                    throw new UnknownResultException("Unknown result received");
             }
         }
 
         /// <summary>
-        /// Clears any text that was written to the display
+        /// Clears any text that was written to the display.
         /// </summary>
+        /// <exception cref="UnknownResultException" />
+        /// <exception cref="CommandNotExecutableException" />
         public void ClearDisplay()
         {
             WriteToDisplay(string.Empty);
         }
 
         /// <summary>
-        /// Changes the display to show the current weight
+        /// Changes the display to show the current weight.
         /// </summary>
+        /// <exception cref="UnknownResultException" />
+        /// <exception cref="CommandNotExecutableException" />
         public void ShowWeightOnDisplay()
         {
             var command = "DW";
@@ -150,18 +177,28 @@ namespace Iot.Device.MettlerToledo
                 case "I":
                     throw new CommandNotExecutableException(command);
                 default:
-                    throw new Exception("Unknown result received");
+                    throw new UnknownResultException("Unknown result received");
             }
         }
 
         /// <summary>
-        /// Have the scale send an event every time the weight is changed
+        /// Have the scale send an event every time the weight is changed.
         /// </summary>
         public void SubscribeToWeightEvents()
         {
             throw new NotImplementedException();
         }
 
+        /// <summary>
+        /// Sends a command over the serial port.
+        /// </summary>
+        /// <param name="command">Command to send</param>
+        /// <returns>Result of the command.</returns>
+        /// <exception cref="InvalidOperationException" />
+        /// <exception cref="TimeoutException">Occurs if the serial port times out while sending or receiving.</exception>
+        /// <exception cref="SyntaxErrorException" />
+        /// <exception cref="TransmissionErrorException" />
+        /// <exception cref="LogicalErrorException" />
         private string[] SendCommandWithResponse(string command)
         {
             _serialPort.WriteLine(command);
@@ -211,135 +248,6 @@ namespace Iot.Device.MettlerToledo
         public void Dispose()
         {
             Close();
-        }
-    }
-
-    /// <summary>
-    /// Weight reading returned by Mettler Toledo scales
-    /// </summary>
-    public class MettlerToledoWeightReading
-    {
-        /// <summary>
-        /// Scale returned this weight value is stable
-        /// </summary>
-        public bool Stable { get; }
-
-        /// <summary>
-        /// Weight returned by the scale
-        /// </summary>
-        public Mass Weight { get; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="MettlerToledoWeightReading"/> class.
-        /// </summary>
-        /// <param name="stable">Whether this value is stable</param>
-        /// <param name="weight">Weight returned by scale</param>
-        public MettlerToledoWeightReading(bool stable, Mass weight) => (Stable, Weight) = (stable, weight);
-    }
-
-    /// <summary>
-    /// A generic Mettler Toledo exception
-    /// </summary>
-    public abstract class MettlerToledoException : System.Exception
-    {
-        /// <summary>
-        /// The command that triggered this exception
-        /// </summary>
-        public string Command { get; set; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="MettlerToledoException"/> class.
-        /// </summary>
-        /// <param name="command">Command that triggered this exception</param>
-        public MettlerToledoException(string command) => Command = command;
-    }
-
-    /// <summary>
-    /// Occurs when a scale returns a Syntax Error.
-    /// </summary>
-    public class SyntaxErrorException : MettlerToledoException
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="SyntaxErrorException"/> class.
-        /// </summary>
-        /// <param name="command">Command that triggered this exception</param>
-        public SyntaxErrorException(string command)
-            : base(command)
-        {
-        }
-    }
-
-    /// <summary>
-    /// Occurs when a scale returns a Transmission Error.
-    /// </summary>
-    public class TransmissionErrorException : MettlerToledoException
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TransmissionErrorException"/> class.
-        /// </summary>
-        /// <param name="command">Command that triggered this exception</param>
-        public TransmissionErrorException(string command)
-            : base(command)
-        {
-        }
-    }
-
-    /// <summary>
-    /// Occurs when a scale returns a Logical Error.
-    /// </summary>
-    public class LogicalErrorException : MettlerToledoException
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="LogicalErrorException"/> class.
-        /// </summary>
-        /// <param name="command">Command that triggered this exception</param>
-        public LogicalErrorException(string command)
-            : base(command)
-        {
-        }
-    }
-
-    /// <summary>
-    /// Occurs when a scale returns a Command Not Executable error. This is typically because the scale is currently running a different command or a timeout occurred.
-    /// </summary>
-    public class CommandNotExecutableException : MettlerToledoException
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="CommandNotExecutableException"/> class.
-        /// </summary>
-        /// <param name="command">Command that triggered this exception</param>
-        public CommandNotExecutableException(string command)
-            : base(command)
-        {
-        }
-    }
-
-    /// <summary>
-    /// Occurs when a scale returns a Balance Overload or Balance Underload error.
-    /// </summary>
-    public class BalanceOutOfRangeException : MettlerToledoException
-    {
-        /// <summary>
-        /// Scale returned it is overloaded.
-        /// </summary>
-        public bool IsOverloaded { get; }
-
-        /// <summary>
-        /// Scale returned it is underloaded.
-        /// </summary>
-        public bool IsUnderloaded { get; }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="BalanceOutOfRangeException"/> class.
-        /// </summary>
-        /// <param name="_overloaded">Scale returned overloaded</param>
-        /// <param name="_underloaded">Scale returned underloaded</param>
-        /// <param name="command">Command that triggered this exception</param>
-        public BalanceOutOfRangeException(bool _overloaded, bool _underloaded, string command)
-            : base(command)
-        {
-            IsOverloaded = _overloaded;
-            IsUnderloaded = _underloaded;
         }
     }
 }
